fix(domain): change salary from double to BigDecimal to avoid precision loss
- Salary value object: double -> BigDecimal with compareTo() for validation
- CreateJobCommand: salaryMin/salaryMax now accept BigDecimal
- JobEntity: add precision=15, scale=2 to @Column for DECIMAL mapping
- V7 migration: alter jobs table columns to DECIMAL(15,2)

+++ src/main/java/com/ITJobsBackend/jobs/application/usecases/createjob/CreateJobCommand.java
@@ -1,12 +1,14 @@
 package com.ITJobsBackend.jobs.application.usecases.createjob;
 
+import java.math.BigDecimal;
+
 public record CreateJobCommand(
     String title,
     String description,
     String company,
     String location,
-    double salaryMin,
-    double salaryMax,
+    BigDecimal salaryMin,
+    BigDecimal salaryMax,
     String currency,
     String employmentType
 ) {}
